adding command getiframe from command.js and improving file division
